Renamed struct handler classes and related tests
